% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bootStrap.R
\name{bootStrap}
\alias{bootStrap}
\title{Boostrap function.}
\usage{
bootStrap(
  cluster = "IDCLUST50",
  inFile = system.file("data", "data.dbf", package = "dataDriven"),
  incProb = "pj",
  depVar = "fj",
  varsToRemove = c("ID", "Y", cluster),
  coordinates = c("x", "y"),
  iterations = 100,
  outDir = system.file("data", package = "dataDriven")
)
}
\arguments{
\item{inFile}{The path for the input dbf file}

\item{incProb}{The name of the probabiliy of inclusion column}

\item{depVar}{The name of the dependent variable column}

\item{varsToRemove}{The name of the columns to do not consider in the analysis}

\item{coordinates}{The number of bootstrap iterations}

\item{outDir}{The directory to save the output shapefile}
}
\value{
The output is the predicted map with the error map associated
}
\description{
This function applies to bootstrap to construct the error map
}
\examples{
s <- bootStrap()

}
\keyword{Akaike,}
\keyword{harmonization}
\keyword{selection,}
\keyword{variables}
