% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datadrivensecurity.R
\name{calculatePercentOfAxis2100NC}
\alias{calculatePercentOfAxis2100NC}
\title{Get percentage of AXIS 2100 Network cameras}
\usage{
calculatePercentOfAxis2100NC(shodan_cves_join)
}
\arguments{
\item{shodan_cves_join, }{joined information of shodan and CVEs data}
}
\value{
axis.percent, percentage of AXIS 2100 Network Cameras
}
\description{
Get percentage of AXIS 2100 Network cameras
}
