% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/completeTies.R
\name{completeTies}
\alias{completeTies}
\title{Implements complete case methodology}
\usage{
completeTies(elements, pattern)
}
\arguments{
\item{elements}{Sub-sets of elements of the time series with dimension d of delay t.}

\item{pattern}{Ordinal patterns formed from the time series with dimension d of delay t.}
}
\value{
The distribution of probability without considering the patterns formed of repeated elements
}
\description{
Calculates the probability distribution using the complete case methodology, that is, using only the patterns formed without repeated elements.
}
\author{
Eduarda Chagas

set.seed(1234567890, kind = "Mersenne-Twister")
x <- runif(110000)
d <- 3
del <- 1
elem <- formationPattern(x = x, dimension = d, delay = del, option = 0)
pat <- formationPattern(x = x, dimension = d, delay = del, option = 1)
completeTies(elements, pattern)
}
