% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datadrivensecurity.R
\name{createFactorsForCVEData}
\alias{createFactorsForCVEData}
\title{Create factor for CVE dataframe}
\usage{
createFactorsForCVEData(df)
}
\arguments{
\item{db, }{data frame with CVEs}
}
\description{
Create factor for CVE dataframe
}
