% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datadrivensecurity.R
\name{createFactorsForShodanData}
\alias{createFactorsForShodanData}
\title{Create factor for Shodan results}
\usage{
createFactorsForShodanData(df)
}
\arguments{
\item{df, }{data frame with Shodan results}
}
\value{
df, data frame with factors for model, vendor, country_name and country_code
}
\description{
Create factor for Shodan results
}
