% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataDriven.R
\name{dataDriven}
\alias{dataDriven}
\title{Calculates Data-Driven Imputation of a Time Series}
\usage{
dataDriven(x, dimension, delay)
}
\arguments{
\item{x}{A numeric vector (e.g. a time series)}

\item{dimension}{Dimension size of ordinal patterns}

\item{delay}{Size of the delay of ordinal patterns}
}
\value{
Numerical vector corresponding to probability distribution with data-driven imputation.
}
\description{
Returns the numeric vector supplied with the given decimal number.
}
\author{
Eduarda Chagas

set.seed(1234567890, kind = "Mersenne-Twister")
x <- runif(110000)
d <- 3
del <- 1
dataDriven(x = x, dimension = d, delay = del)
}
