% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/definePatterns.R
\name{definePatterns}
\alias{definePatterns}
\title{Calculates all ordinal patterns with a given dimension}
\usage{
definePatterns(d)
}
\arguments{
\item{d}{Dimension size of ordinal patterns}
}
\value{
A numerical matrix with the patterns
}
\description{
Returns the patterns with dimension d
}
\author{
Eduarda Chagas

d <- 3
definePatterns(d = d)
}
