% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datadrivensecurity.R
\name{downloadCVEData}
\alias{downloadCVEData}
\title{Download CVE data}
\usage{
downloadCVEData()
}
\value{
CVE data frame
}
\description{
Download CVE data information from https://github.com/r-net-tools/security.datasets/raw/master/net.security/sysdata.rda.
}
