% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/equalitiesValues.R
\name{equalitiesValues}
\alias{equalitiesValues}
\title{Calculates the percentage of repeated elements in a time series}
\usage{
equalitiesValues(x)
}
\arguments{
\item{x}{A numeric vector (e.g. a time series)}
}
\value{
Percentage of repeated elements of a time series.
}
\description{
Analyzes and calculates the percentage of how many repeated elements a time series has.
}
\author{
Eduarda Chagas

set.seed(1234567890, kind = "Mersenne-Twister")
x <- runif(110000)
equalitiesValues(x = x)
}
