% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formationPattern.R
\name{formationPattern}
\alias{formationPattern}
\title{Form the ordinal patterns of a time series}
\usage{
formationPattern(serie, dimension, delay, option)
}
\arguments{
\item{serie}{A numeric vector (e.g. a time series)}

\item{dimension}{Dimension size of ordinal patterns}

\item{delay}{Size of the delay of ordinal patterns}

\item{option}{Determines the data set to be returned. The parameter must be 0 for ordinal patterns formed or different from 0 to return the subsets of formed elements.}
}
\value{
A numerical vector containing ordinal patterns or subsets of formed elements.
}
\description{
Divides the time series into subsets with a certain size and delay and calculates its ordinal patterns.
}
\author{
Eduarda Chagas

set.seed(1234567890, kind = "Mersenne-Twister")
x <- runif(110000)
d <- 3
del <- 1
opt <- 0
formationPattern(serie = x, dimension = d, delay = del, option = opt)
}
