% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datadrivensecurity.R
\name{getCameraSoftware}
\alias{getCameraSoftware}
\title{Apply parsing functions depending on the camera facturer}
\usage{
getCameraSoftware(software_info, shodan_query)
}
\arguments{
\item{software_info, }{list with shodan searches}

\item{shodan_query, }{value of the query used when searching with Shodan}
}
\value{
software_info data frame with extra information of vendor, model and version
}
\description{
Apply parsing functions depending on the camera facturer
}
