% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datadrivensecurity.R
\name{getJoinedInfoWithNoNa}
\alias{getJoinedInfoWithNoNa}
\title{Remove Na values}
\usage{
getJoinedInfoWithNoNa(df)
}
\arguments{
\item{df, }{data frame with Na values}
}
\value{
df, data frame without Na values
}
\description{
Remove Na values
}
