% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datadrivensecurity.R
\name{getVulnerableCamerasCVEs}
\alias{getVulnerableCamerasCVEs}
\title{Get data frame with CVEs information related to cameras vulnerabilities}
\usage{
getVulnerableCamerasCVEs(cves)
}
\arguments{
\item{cves, }{entire list of CVEs}
}
\value{
cve_cameras_data, data frame with CVEs
}
\description{
Get data frame with CVEs information related to cameras vulnerabilities
}
