% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/harmon.R
\name{harmon}
\alias{harmon}
\title{Harmonized map construction.}
\usage{
harmon(inFile, incProb, depVar, varsToRemove, coordinates, AK)
}
\arguments{
\item{inFile}{The path for the input dbf file}

\item{incProb}{The name of the probabiliy of inclusion column}

\item{depVar}{The name of the dependent variable column}

\item{varsToRemove}{The name of the columns to do not consider in the analysis}

\item{coordinates}{}
}
\value{
The output is a list with length 4:
(1) v_aux: matrice di k+1 elementi (colonna 1:nome v. ausiliarie, colonna 2: 1 se la v. ausiliaria ? selezionata e 0 altrimenti)
(2) coeff: numero delle v. ausiliarie selezionalte
(3) founded_a: valore di alpha scelto dalla LOOCV
(4) estim: coordinate delle unit? areali e relativo valore armonizzato della variabile di interesse. Da estim si può plottare la mappa armonizzata
}
\description{
This function constructs the harmonized map
}
\examples{
h <- harmon(
  inFile = "data.dbf",
  incProb = "pj",
  depVar = "fj",
  varsToRemove =  c("ID", "IDCLUST50", "Y"),
  coordinates = c("x", "y"),
  AK = AK
)
}
\keyword{Akaike,}
\keyword{harmonization}
\keyword{selection,}
\keyword{variables}
