% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datadrivensecurity.R
\name{leftJoinShodanWithCVEsInfo}
\alias{leftJoinShodanWithCVEsInfo}
\title{Join Shodan results with CVEs information}
\usage{
leftJoinShodanWithCVEsInfo(shodan_cameras_data, cve_cameras_data)
}
\arguments{
\item{shodan_cameras_data, }{data frame with Shodan results}

\item{cve_cameras_data, }{data frame with CVEs information}
}
\value{
schodan_cves_join, data frame with joined information
}
\description{
Join Shodan results with CVEs information
}
