% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Main.R
\name{Main}
\alias{Main}
\title{Main function.}
\usage{
Main(
  inFile = system.file("data", "data.dbf", package = "dataDriven"),
  incProb = "pj",
  depVar = "fj",
  varsToRemove = c("ID", "IDCLUST50", "Y"),
  coordinates = c("x", "y")
)
}
\arguments{
\item{inFile}{The path for the input dbf file}
}
\value{
The output is a list with length 4:
(1) v_aux: matrice di k+1 elementi (colonna 1:nome v. ausiliarie, colonna 2: 1 se la v. ausiliaria ? selezionata e 0 altrimenti)
(2) coeff: numero delle v. ausiliarie selezionalte
(3) founded_a: valore di alpha scelto dalla LOOCV
(4) estim: coordinate delle unit? areali e relativo valore armonizzato della variabile di interesse. Da estim si può plottare la mappa armonizzata
}
\description{
This function constructs the harmonized map
}
\examples{
h <- Main()
}
\keyword{Akaike,}
\keyword{harmonization}
\keyword{selection,}
\keyword{variables}
