% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/permute.R
\name{myPermute}
\alias{myPermute}
\title{Calculates all possible combinations of a set by exchanging only a subset of elements}
\usage{
myPermute(vector, init, end)
}
\arguments{
\item{vector}{A numeric vector}

\item{init}{Beginning of subset}

\item{end}{End of subset}
}
\value{
Numerical matrix with permutations performed.
}
\description{
Returns a numeric array containing all possible permutations of a vector by varying only a given subset.
}
\author{
Eduarda Chagas

x <- c(1:10)
i <- 2
e <- 6
mypermute(vector = x, init = i, end = e)
}
