% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datadrivensecurity.R
\name{parseDlinkCamera}
\alias{parseDlinkCamera}
\title{Parse D-Link camera information got from Shodan searches}
\usage{
parseDlinkCamera(software_info, regex_camera_search)
}
\arguments{
\item{software_info, }{list with shodan searches}

\item{regex_camera_search, }{regex experssion used to parse AXIS cameras information}
}
\value{
software_info data frame with extra information for vendor, model and version
}
\description{
Parse D-Link camera information got from Shodan searches
}
