% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datadrivensecurity.R
\name{parseSonyCamera}
\alias{parseSonyCamera}
\title{Parse Sony camera information got from Shodan searches}
\usage{
parseSonyCamera(software_info, regex_camera_search)
}
\arguments{
\item{software_info, }{list with shodan searches}

\item{regex_camera_search, }{regex experssion used to parse Sony cameras information}
}
\value{
software_info data frame with extra information for vendor, model and version
}
\description{
Parse Sony camera information got from Shodan searches
}
