% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/precision.R
\name{precision}
\alias{precision}
\title{Determines a precise number of a set of numbers}
\usage{
precision(x, y)
}
\arguments{
\item{x}{A numeric vector (e.g. a time series)}

\item{y}{Number of decimal places}
}
\value{
The vector \code{x} com \code{y} decimal places.
}
\description{
Returns the numeric vector supplied with the given decimal number.
}
\author{
Eduarda Chagas

set.seed(1234567890, kind = "Mersenne-Twister")
x <- runif(110000)
y <- 4
precision(x = x, y = y)
}
