% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datadrivensecurity.R
\name{searchCamerasWithShodan}
\alias{searchCamerasWithShodan}
\title{Search cameras with Shodan}
\usage{
searchCamerasWithShodan(shodan_query)
}
\arguments{
\item{shodan_query, }{query used for searching information}
}
\description{
Search cameras with Shodan
}
