% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shannonNormalized.R
\name{shannonNormalized}
\alias{shannonNormalized}
\title{Calculates the normalized Shannon entropy}
\usage{
shannonNormalized(p)
}
\arguments{
\item{p}{A numerical vector containing an ordinal pattern distribution.}
}
\value{
Entropy of Shannon.
}
\description{
Returns the normalized Shannon entropy of a probability distribution
}
\author{
Eduarda Chagas

set.seed(1234567890, kind = "Mersenne-Twister")
x <- runif(110000)
d <- 3
del <- 1
prob = dataDriven(x = x, dimension = d, delay = del)
shannonNormalized(p = prob)
}
