% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/selectVariables.R
\name{variableSelection}
\alias{variableSelection}
\title{Akaike variable selection.}
\usage{
variableSelection(inFile, incProb, depVar, varsToRemove, coordinates)
}
\arguments{
\item{inFile}{The path for the input dbf file}

\item{incProb}{The name of the probabiliy of inclusion column}

\item{depVar}{The name of the dependent variable column}

\item{varsToRemove}{The name of the columns to do not consider in the analysis}

\item{coordinates}{}
}
\value{
The output is a list with length 3:
(1) v_aux: valori delle variabili ausiliare selezionate nelle unit? campionate
(2) v_aux_U: valori delle variabili ausiliare selezionate nelle unit? non campionate
(3) varprov: vettore di k+1 elementi (il primo ? 1 di default, gli altri sono 1 se la v. ausiliaria ? selezionata e 0 altrimenti)
}
\description{
This function performs the variables selection.
}
\examples{
AK <- variableSelection(
  inFile = "data.dbf",
  incProb = "pj",
  depVar = "fj",
  varsToRemove =  c("ID", "IDCLUST50", "Y"),
  coordinates = c("x", "y")
)
}
\keyword{Akaike,}
\keyword{selection}
\keyword{variables}
