% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bootStrap.R
\name{bootStrap}
\alias{bootStrap}
\title{Boostrap function.}
\usage{
bootStrap(
  cluster = "IDCLUST",
  inFile = system.file("data", "data.dbf", package = "dataDriven"),
  depVar = "fj",
  varsToRemove = c("ID", "Y"),
  coordinates = c("x", "y"),
  resampling = 1000,
  outDir = system.file("data", package = "dataDriven")
)
}
\arguments{
\item{cluster}{The ID strata of the population spatial units. (If simple random sampling is adopted, set cluster = 1)}

\item{inFile}{The path for the input dbf file. By default, it is automatically selected a test file provided within the package. To see it use system.file("data", "data.dbf", package = "dataDriven")}

\item{depVar}{The name of the interest variable column}

\item{varsToRemove}{The name of the columns to do not consider in the analysis (if there are any)}

\item{coordinates}{The spatial coordinates of the population areal units}

\item{resampling}{The number of bootstrap resampling}

\item{outDir}{The directory to save the output shapefile}
}
\value{
The output is the predicted map with the associated error map
}
\description{
This function applies to bootstrap to construct the error map
}
\examples{
s <- bootStrap()

}
\keyword{Akaike,}
\keyword{harmonization}
\keyword{selection,}
\keyword{variables}
