% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/harmon.R
\name{harmon}
\alias{harmon}
\title{Harmonized map construction.}
\usage{
harmon(db, depVar, varsToRemove, coordinates, AK)
}
\arguments{
\item{depVar}{The name of the interest variable column}

\item{varsToRemove}{The name of the columns to do not consider in the analysis (if there are any)}

\item{coordinates}{The spatial coordinates of the population areal units}

\item{inFile}{The path for the input dbf file}

\item{incProb}{The name of first-order inclusion probability column}
}
\value{
The output is a list with length 4:
(2) coeff: number of auxiliary variables selected with the Akaike-type criterion.
(3) founded_a: alpha choosen by the LOOCV
(4) estim: spatial coordinates of the areal units and the corresponding harmonized values of the interest attribute.
It can be used to plot the harmonized map.
}
\description{
This function constructs the harmonized map
}
\examples{
h <- harmon(
  inFile = "data.dbf",
  incProb = "pj",
  depVar = "fj",
  varsToRemove =  c("ID", "IDCLUST50", "Y"),
  coordinates = c("x", "y"),
  AK = AK
)
}
\keyword{Akaike,}
\keyword{harmonization}
\keyword{selection,}
\keyword{variables}
