% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/selectVariables.R
\name{variableSelection}
\alias{variableSelection}
\title{Akaike variable selection.}
\usage{
variableSelection(inFile, depVar, cluster, varsToRemove, coordinates)
}
\arguments{
\item{inFile}{The path for the input dbf file}

\item{depVar}{The name of the interest variable column}

\item{varsToRemove}{The name of the columns to do not consider in the analysis (if there are any)}

\item{coordinates}{The spatial coordinates of the population areal units}
}
\value{
The output is a list with length 3:
(1) v_aux: values of auxiliary variables selected in the sampled spatial units
(2) v_aux_U: values of auxiliary variables selected in the sampled spatial units
(3) varprov: vector of k+1 elements (the first is = 1 for default, the others are = 1 if the auxiliary variable is selected and 0 otherwise)
}
\description{
This function performs the variables selection.
}
\examples{
AK <- variableSelection(
  inFile = "data.dbf",
  depVar = "fj",
  varsToRemove =  c("ID", "IDCLUST50", "Y"),
  coordinates = c("x", "y")
)
}
\keyword{Akaike,}
\keyword{selection}
\keyword{variables}
